requirements.txt
flask==2.2.3
pillow==9.5.0
numpy==1.24.3
google-generativeai==0.3.2